Split threading in multiple files (similar to TBB) and update includes.
