+++ components/qslim/wmtk/components/qslim/QSlimMesh.h
@@ -8,7 +8,7 @@
 // clang-format off
 #include <wmtk/utils/DisableWarnings.hpp>
 #include <igl/write_triangle_mesh.h>
-#include <wmtk/threading/Concurrency.hpp>
+#include <wmtk/threading/enumerable_thread_specific.hpp>
 #include <wmtk/envelope/Envelope.hpp>
 #include <wmtk/utils/EnableWarnings.hpp>
 // clang-format on
