+++ components/simwild/wmtk/components/simwild/SimWildMeshTri.cpp
@@ -1,5 +1,8 @@
 #include "SimWildMeshTri.hpp"
-#include <wmtk/threading/Concurrency.hpp>
+#include <wmtk/threading/enumerable_thread_specific.hpp>
+#include <wmtk/threading/parallel_for.hpp>
+#include <wmtk/threading/parallel_sort.hpp>
+#include <wmtk/threading/task_arena.hpp>
 
 #include <igl/Timer.h>
 #include <igl/is_edge_manifold.h>
