+++ components/tetwild/wmtk/components/tetwild/TetWildMesh.h
@@ -5,14 +5,19 @@
 #include <wmtk/utils/Morton.h>
 #include <wmtk/utils/PartitionMesh.h>
 #include <wmtk/envelope/Envelope.hpp>
+#include <wmtk/threading/concurrent_map.hpp>
+#include <wmtk/threading/enumerable_thread_specific.hpp>
+#include <wmtk/threading/parallel_for.hpp>
+#include <wmtk/threading/parallel_sort.hpp>
+#include <wmtk/threading/task_arena.hpp>
+
 #include "Parameters.h"
 
 // clang-format off
 #include <wmtk/utils/DisableWarnings.hpp>
 #include <fastenvelope/FastEnvelope.h>
-#include <wmtk/threading/Concurrency.hpp>
-#include <wmtk/utils/EnableWarnings.hpp>
 #include <VolumeRemesher/embed.h>
+#include <wmtk/utils/EnableWarnings.hpp>
 // clang-format on
 
 #include <igl/remove_unreferenced.h>
